fix the invalid double free in tdf8532 i2c read error
tdf8532_single_read() is defined as return uint8_t, the error
ret is defined as int, when it run into the error case,
the tdf8532_single_read() won't return the error, it will
cause the double free.

Tracked-On: https://jira01.devtools.intel.com/browse/OAM-69442
Reviewed-on: https://git-gar-1.devtools.intel.com/gerrit/31907
Tested-by: Rojewski, Cezary <cezary.rojewski@intel.com>

+++ sound/soc/codecs/tdf8532.c
@@ -83,7 +83,7 @@ static int __tdf8532_single_write(struct tdf8532_priv *dev_data,
 }
 
 
-static uint8_t tdf8532_read_wait_ack(struct tdf8532_priv *dev_data,
+static int tdf8532_read_wait_ack(struct tdf8532_priv *dev_data,
 						unsigned long timeout)
 {
 	uint8_t ack_repl[HEADER_SIZE] = {0, 0, 0};
